Move notifications into dedicated app
