+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project are documented in this file.
 ### Added
 - Added a CodeQL analysis workflow for backend Python, frontend JavaScript/TypeScript, and GitHub Actions workflow scanning.
 - Added recipient-scoped in-app notifications with API endpoints for listing and marking notifications read.
+- Added a dedicated backend `notifications` app to own notification models, API routes, admin, services, and tests.
 - Added app bar notification badge and popover UI for shared board activity.
 - Added notification generation for collaborator adds, shared board list creation, note creation, and note updates.
 ### Fixed
