+++ backend/app/urls.py
@@ -21,5 +21,6 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/", include("notes.urls")),
+    path("api/", include("notifications.urls")),
     path("auth/", include("authentication.urls")),
 ]
