+++ backend/notifications/urls.py
@@ -1,0 +1,11 @@
+from django.urls import include, path
+from rest_framework.routers import DefaultRouter
+
+from .views import NotificationViewSet
+
+router = DefaultRouter()
+router.register(r"notifications", NotificationViewSet, basename="notification")
+
+urlpatterns = [
+    path("", include(router.urls)),
+]
